Move get_cache_entry_from_duckdb_index_job to libapi
